requirements.txt
appdirs==1.4.4
asgiref==3.2.10
astroid==2.4.2
attrs==19.3.0
black==20.8b1
celery==4.3.0
certifi==2020.6.20
click==7.1.2
coverage==5.2
Django==3.0.8
django-cors-headers==3.7.0
django-extensions==3.0.2
django-redis==4.11.0
django-ses==2.2.0
djangorestframework==3.11.0
djangorestframework-simplejwt==4.7.1
factory-boy==3.2.0
flake8==3.8.3
fluent-logger==0.9.6
names==0.3.0
psycopg2>=2.8.6,<2.9
psycopg2-binary==2.8.5
py==1.9.0
pycodestyle==2.6.0
pyflakes==2.2.0
pylint==2.5.3
pylint-django==2.0.15
pylint-plugin-utils==0.6
pyparsing==2.4.7
pytest==5.4.3
pytest-black==0.3.10
pytest-cov==2.10.0
pytest-django==3.8.0
pytest-flake8==1.0.6
pytest-factoryboy==2.1.0
pytest-freezegun==0.4.1
pytest-isort==1.1.0
pytest-subtests==0.5.0
python-dateutil==2.8.1
pytz==2020.1
uWSGI>=2.0.19.1,<2.1
redis==3.5.3
safety==1.10.3